Changed locations of the app.py, config.py and models.py, also revamped SQLite3 in login and register
