--- app/__init__.py
@@ -1,4 +1,0 @@
-from flask import Flask
-app = Flask(__name__)
-
-from app import routes
