+++ app/userauth/__init__.py
@@ -1,14 +1,23 @@
+from app import app
 from flask import Flask
+from flask_login import LoginManager
 from config import Config
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate
-from .routes import routes
-from .models import models
 
 
-from app import routes
+
 app = Flask(__name__)
 app.config.from_object(Config)
 db = SQLAlchemy(app)
 migrate = Migrate(app, db)
+
 from app import routes, models
+from models import User
+login_manager = LoginManager()
+login_manager.init_app(app)
+
+@login_manager.user_loader
+def load_user(user_id):
+    user = User.query.get(user_id)
+    return user
