+++ app/userauth/forms.py
@@ -1,5 +1,7 @@
-import flask
-from flask_tf import FlaskForm
+import flask 
+from flask import Flask
+from flask_wtf import FlaskForm
+from flask_sqlalchemy import SQLAlchemy
 from wtforms import IntegerField, StringField, BooleanField, SubmitField
 from wtforms.validators import DataRequired
 
@@ -13,3 +15,5 @@ class RegisterForm(FlaskForm):
     username = StringField('Username', validators=[DataRequired])
     password = StringField('Password', validators=[DataRequired])
     submit = SubmitField('Register')
+
+
